refactor(www/audio): move all AudioQueue GraphQL hooks into a single file
